+++ cmd/main.go
@@ -11,25 +11,19 @@ import (
 
 var db *sql.DB
 
-//フォーム入力された情報
-type UserInput struct {
-	UserId string `json:"userId"`
-	Pass   string `json:"userPass"`
-}
-
-//DB情報
-type DBUser struct {
-	UserId     string
-	Pass       string
-	MentorFlag int
-}
-
 func main() {
+	db, err := sql.Open("mysql", "moxyo3:moxyo3@/test_db")
+	if err != nil {
+		log.Print("DB接続エラー")
+	}
+
 	http.Handle("/", http.StripPrefix("/", http.FileServer(http.Dir("../src"))))
 	http.HandleFunc("/loginCheck", loginCheck)
 
 	log.Println("start http server :8080")
 	log.Fatal(http.ListenAndServe(":8080", nil))
+
+	defer db.Close()
 }
 
 func loginCheck(w http.ResponseWriter, r *http.Request) {
@@ -41,43 +35,28 @@ func loginCheck(w http.ResponseWriter, r *http.Request) {
 	}
 	log.Print(userInput)
 
-	//db処理
-	db, err := sql.Open("mysql", "moxyo3:moxyo3@/test_db")
-	if err != nil {
-		panic(err.Error())
-	}
-	defer db.Close()
-
 	//dbから一致したレコードを格納
 	var user DBUser = DBUser{}
 
 	//userid,passが一致するものを検索
 	if err := db.QueryRow("SELECT * FROM users WHERE user_id=? AND pass=?", (&userInput).UserId, (&userInput).Pass).Scan(&user.UserId, &user.Pass, &user.MentorFlag); err != nil {
-		switch {
-		case err == sql.ErrNoRows:
-			log.Fatal("アカウントが存在しません")
-		case err != nil:
-			log.Fatal(err)
+		if err == sql.ErrNoRows {
+			log.Print("アカウントが存在しません")
+		} else {
+			log.Print(err)
 		}
+		log.Print(user)
 	}
-	log.Print(user)
 }
 
-func createUser(w http.ResponseWriter, r *http.Request) {
+/*func createUser(w http.ResponseWriter, r *http.Request) {
 	var newUser UserInput
 	if err := json.NewDecoder(r.Body).Decode(&newUser); err != nil {
 		http.Error(w, err.Error(), 500)
 		return
 	}
-	//db処理
-	db, err := sql.Open("mysql", "moxyo3:moxyo3@/test_db")
-	if err != nil {
-		panic(err.Error())
-	}
-	defer db.Close()
-
 	//dbのレコードを格納
 	//var user DBUser = DBUser{}
 
 	//user_idが重複していないかチェック
-}
+}*/
